requirements.txt
ipykernel
openai>=1.0.0
python-dotenv
yfinance
pandas
matplotlib
Pillow
pyaudio
SpeechRecognition
sounddevice
playsound==1.2.2